requirements.txt
Flask==2.2.3
Flask-Bootstrap4==4.0.2
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
email-validator==2.0.0.post2
WTForms-Alchemy==0.18.0
pre-commit==3.2.2